backend/requirements.txt
fastapi
uvicorn
tqdm
numpy
python-dotenv
requests
beautifulsoup4
google-generativeai
